### 2024-04-23 #6665 @editorialbot
[PRE REVIEW]: One4All: An Open Source Portal to Validate and Share Microplastics Data and Beyond
**Submitting author:** @hsherrod2019 (<a href="http://orcid.org/0009-0001-0497-8693">Hannah Sherrod</a>)
**Repository:** https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All
**Branch with paper.md** (empty if default branch):
**Version:** v0.4
**Reviewers:** @hahsan1, @@stephpenn1, @stephpenn1
