### 2024-04-24 690c009
added MIT license; modified readme to reflect this
