+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   app: app.html
   package: package.html
-last_built: 2024-04-24T16:07Z
+last_built: 2024-04-24T16:09Z
 urls:
   reference: https://moore-institute-4-plastic-pollution-res.github.io/One4All/reference
   article: https://moore-institute-4-plastic-pollution-res.github.io/One4All/articles
