### 2024-04-24 #6665 @hsherrod2019
> In the meantime @hsherrod2019, could you assign an open source software license to your codebase? See list here: https://opensource.org/licenses

Yes, I just assigned the MIT license. Thank you! @crvernon
