+++ README.md
@@ -31,7 +31,7 @@ remotes::install_github("Moore-Institute-4-Plastic-Pollution-Res/One4All")
 ```
 
 ### Access the validator tool below:
-- Click here for the web app: [openanalysis.org/one4all/](https://openanalysis.org/one4all/).
+- Click here for the web app: [openanalysis.org/microplastic_data_portal/](https://openanalysis.org/microplastic_data_portal/).
 
 - The code for the app is in [code/validator](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/tree/main/code/validator)
 
