### 2024-04-24 4ac6778
updated readme in validator
