+++ code/validator/README.md
@@ -17,7 +17,7 @@ Funded by the National Renewable Energy Laboratory and Possibility Lab, maintain
   library(One4All)
 ```
 
--   or use the web app: [openanalysis.org/one4all/](https://openanalysis.org/microplastic_data_portal/)
+-   or use the web app: [openanalysis.org/microplastic_data_portal/](https://openanalysis.org/microplastic_data_portal/)
 
 -   The code for the app is in [ui.R](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/code/validator/ui.R), [server.R](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/code/validator/server.R), and [global.R](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/code/validator/global.R)
 
