### 2024-05-02 #6715 @kanishkan91
@hsherrod2019 A couple minor adjustments I will request as the review progresses-

1. Your current word count is on the higher side. We usually recommend somewhere along ~1000 words for a JOSS publication. Looking at the paper, the "workflow overview" section is a bit too long. I would recommend you move most of this to the documentation, vignettes (if it does not exist there already). And maybe just limit to 1 figure which provides an overview.
2. Take a look at the DOI tests above in this thread. There are some warnings that popped up about missing DOIs. Could you take a look and confirm no DOIs are missing?

Other than that reviewers will keep you posted. Let me know what you think. Thanks!
