### 2024-05-03 #6715 @kanishkan91
@hsherrod2019  I can reach out to the EiC to find out how to proceed. Though is it that you just want to edit the paper (which includes adding more co-authors) or do you envision substantial code updates? If its just the former, I would recommend continuing with the review. Let me know what you think
