### 2024-05-03 #6715 @kanishkan91
@hsherrod2019 Also based on consultation with the JOSS EiC, are you confident you can complete all changes within a month? If so, I will pause this review and we can revisit in a month. Else you would need to withdraw the manuscript  and then resubmit on a later date. Let me know which works for you.

Also tagging the EiC in charge here- @crvernon .
