### 2024-05-03 #6715 @wincowgerDEV
@kanishkan91 and @crvernon, Thank you for your flexibility here. I am @hsherrod2019's mentor and coauthor on the paper so am responding just to apologize for the last-minute changes, this is on me, not her. We are planning on adding coauthors from another tightly linked code project already mentioned in the manuscript (microplastic data visualization tool and microplastic image explorer). I think in an ideal world, these code bases would also be reviewed. Do they need to be included in the same repo as the current submission? If so that will take some reorganizing but shouldn't take more than a month.
