### 2024-05-03 #6715 @kanishkan91
@wincowgerDEV From the description above, this sounds like a more detailed software package including several repositories? (Correct me if I'm wrong). Ideally a JOSS paper is limited to a piece of software that can be installed from a single repository. And if these are independent tools that are on separate repos, then that would require separate papers.

Could you clarify if these are separate tools on separate repos that are part of an ecosystem?
