### 2024-05-03 #6715 @wincowgerDEV
@kanishkan91, thanks for the clarification, we were wondering that too. The repo we currently have under review is an r package and app that is basically the foundation for the other two apps. We don't have an intent to create other r packages for those apps because all foundational code will be put into this package. Yes these are separate tools on separate repos that are part of an ecosystem, the big vision being described in this paper to share high-quality microplastic data and have it be easily visualized and queried.
