### 2024-05-03 #6715 @kanishkan91
@wincowgerDEV Thanks! that clarifies more.

My take would be that if you can integrate all code into a single software package that can be invoked from a single repo and rewrite the paper for that single software package, that would be the most ideal for the paper review.

I would still need to check with the EiC (@crvernon ) since this may mean they would need to re-review the paper in that case as well.

But my take (subject to what the EiC says) is that if this can be accomplished in the next month, I can pause this review. Else you can withdraw it and submit at your convenience. You can always refer to this issu number when you resubmit, so JOSS is aware of the paper's history.
