### 2024-05-07 5885d14
added image explorer vignette; updated validator vignette
