+++ pkgdown.yml
@@ -3,8 +3,9 @@ pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
   app: app.html
+  imageexplorer: imageexplorer.html
   package: package.html
-last_built: 2024-04-26T21:45Z
+last_built: 2024-05-07T17:43Z
 urls:
   reference: https://moore-institute-4-plastic-pollution-res.github.io/One4All/reference
   article: https://moore-institute-4-plastic-pollution-res.github.io/One4All/articles
