### 2024-05-07 bb16e83
added data visualization pkgdown + images
