+++ vignettes/imageexplorer.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Image Explorer"
+title: "Microplastic Image Explorer"
 author: "Hannah Sherrod, Win Cowger"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
