### 2024-05-07 b2c38f9
updated vignette titles to match indexentry
