+++ vignettes/datavisualization.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Data Visualization"
+title: "Data Visualization Tutorial"
 author: "Hannah Sherrod, Nicholas Leong, Win Cowger"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Image Explorer Tutorial}
+  %\VignetteIndexEntry{Data Visualization Tutorial}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
 ---
