+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   datavisualization: datavisualization.html
   imageexplorer: imageexplorer.html
   package: package.html
-last_built: 2024-05-07T19:47Z
+last_built: 2024-05-07T19:59Z
 urls:
   reference: https://moore-institute-4-plastic-pollution-res.github.io/One4All/reference
   article: https://moore-institute-4-plastic-pollution-res.github.io/One4All/articles
