### 2024-05-08 933c867
update to description (changed CC BY 4.0 to MIT License)
