### 2024-05-08 5cfe18b
modified the run_app function so users can indicate which app to load; (note: switch function from MDP to One4All once other apps are in One4All)
