+++ code/microplastic_image_explorer/data_cleaning.R
@@ -1,0 +1,225 @@
+library(readxl)
+library(googledrive)
+library(dplyr)
+library("jpeg")
+library("tiff")
+library(magick)
+library(data.table)
+
+#Info we want Timestamp	Image File	Researcher Name	Affiliation	Citation	Instrument name	Analysis Date	Polymer-type of particle	Magnification	Color	Morphology	Size of particle	Size dimension
+
+#Convert images to thumbnails
+
+#Cheng ----
+Cheng <- read_xlsx(path = "G:/My Drive/MooreInstitute/Projects/PeoplesLab/Code/Microplastic_Data_Portal/data/microplastic_images/OneDrive_1_4-14-2024/Microplastic Image Explorer Metadata.xlsx") %>%
+    mutate(name = paste0(ImageID, ".jpg"))
+
+files_in_list <- list.files("G:/My Drive/MooreInstitute/Projects/PeoplesLab/Code/Microplastic_Data_Portal/data/microplastic_images/OneDrive_1_4-14-2024/Images", pattern = ".tif", full.names = T)
+files_in_list_names <- list.files("G:/My Drive/MooreInstitute/Projects/PeoplesLab/Code/Microplastic_Data_Portal/data/microplastic_images/OneDrive_1_4-14-2024/Images", pattern = ".tif")
+
+#for(image in 1:length(files_in_list)){
+#    img <- readTIFF(files_in_list[image], native=TRUE)
+#    writeJPEG(img, target = paste0("G:/My Drive/MooreInstitute/Projects/PeoplesLab/Code/Microplastic_Data_Portal/data/microplastic_images/OneDrive_1_4-14-2024/Images/",gsub("\\..*", "", files_in_list_names[image]), ".jpg"), quality = 1)
+#}
+
+
+#Leah ----
+jpegs <- list.files("C:/Users/winco/OneDrive/Documents/Images_MicroplasticImageExplorer/Leah", pattern = ".jpg", recursive = T, full.names = T)
+shortjpegs <- gsub(".jpg", "", gsub(".*/", "", jpegs))
+info <- read.csv("C:/Users/winco/OneDrive/Documents/Images_MicroplasticImageExplorer/tbl_qa_master.csv")
+microscopy <- read.csv("C:/Users/winco/OneDrive/Documents/Images_MicroplasticImageExplorer/tbl_microscopysettings.csv")
+info_clean <- info %>%
+    left_join(microscopy, by = "objectid") %>%
+    mutate(file = paste0(particleid, ".jpg")) %>%
+    mutate(timestamp = NA) %>%
+    mutate(citation = "SCCWRP Interlaboratory Comparison 2020-2022",
+           instrument = NA, 
+           affiliation = "SCCWRP", 
+           researcher = paste0("Lab ", labid.x), 
+           analysis_date = "2020-2022", 
+           dimension = "nominal")
+
+sum(shortjpegs %in% info$particleid)
+cleanedjpgs <- jpegs[shortjpegs %in% info$particleid]
+cleanedshortjpgs <- shortjpegs[shortjpegs %in% info$particleid]
+
+#for(item in 1:length(cleanedjpgs)){
+#    file.copy(from = cleanedjpgs[item], to = paste0("C:/Users/winco/OneDrive/Documents/Images_MicroplasticImageExplorer/LeahClean/", cleanedshortjpgs[item], ".jpg"))
+#}
+
+#drive_deauth()
+#files <- drive_ls(drive_get(as_id("https://drive.google.com/drive/folders/103OUoOpOqxgn06fJA2Rq38SjFgfdejbRcvanC9u2juKqelwmgzrL0f7xI8T9G-_z7r6XbAeb")))
[…]
