### 2024-05-08 9fd33ac
moved apps into inst/apps
