### 2024-05-09 a0fd090
updates to vignettes, to '.Rbuildignore', and to app locations
