### 2024-05-09 6bd92e6
update to run_app function in global.R
