### 2024-05-09 d0c9e2f
added download_all function in global.R and tests.R
