+++ inst/apps/validator/ui.R
@@ -1,0 +1,197 @@
+function(request) {
+    dashboardPage(
+    fullscreen = T,
+    help = T,
+    dashboardHeader(title = config$portal_name,
+                    includeCSS("www/datatable.css")),
+    dashboardSidebar(
+        sidebarUserPanel(
+            name = config$portal_funder_name,
+            image = config$portal_funder_link
+        ),
+        sidebarMenu(
+            id = "sidebarmenu",
+            menuItem(
+                "Uploader",
+                tabName = "validator",
+                icon = icon("upload")
+            ),
+            menuItem(
+                "Downloader",
+                tabName = "downloader",
+                icon = icon("download")
+            ),
+            menuItem(
+                "About",
+                tabName = "about",
+                icon = icon("sliders-h")
+            ),
+            #menuItem(
+            #  "Submission Guidelines",
+            #  tabName = "item4",
+            #  icon = icon("chalkboard")
+            #),
+            menuItem(
+                "Help",
+                tabName = "help",
+                icon = icon("question")
+            )
+        )
+    ),
+    dashboardBody(
+        shinyjs::useShinyjs(),
+        tabItems(
+            tabItem(
+                tabName = "about",
+                box(
+                    title = "Overview",
+                    p("Welcome to the Data Validator webpage. This tool allows you to validate data interactively by uploading a dataset and rules file. To get started, go to the validator tab on the left."),
+                    HTML(paste0('<iframe width="560" height="315" src="',config$overview,'" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>')),
+                    width = 12
+                ),
+                box(
+                    title = "Contribute",
+                    collapsed = F,
+                    p("Join our team to build this tool!"),
+                    HTML(paste0('<a class="btn btn-info" href = "', config$github,'" role = "button" >Github</a>')),
+                    boxLayout(
+                        type = "columns",
+                        lapply(config$contributors, function(x){x})
+                    ),
+                    width = 12
+                )
+            ),
+            tabItem(
+                tabName = "validator",
+                fluidRow(
+                    column(4,
+                               fileInput("file", NULL,
+                                         placeholder = "Start Here",
+                                         buttonLabel = "Upload Data",
+                                         width = "100%",
+                                         multiple = T,
+                                         accept=c("text/csv",
+                                                  "text/comma-separated-values,text/plain",
+                                                  ".xlsx",
+                                                  ".zip")) %>%
+                               popover(
+                               title = "Upload CSV to validate",
+                               content = "This is where you upload the csv, zip, and/or xlsx files file that you want to validate.")
+                    ),
+                    column(4,
+                           if(!isTruthy(config$rules_to_use)){
+                               popover(
+                                   fileInput("file_rules", NULL,
+                                             placeholder = ".csv",
+                                             buttonLabel = "Rules...",
+                                             width = "100%",
+                                             accept=c("text/csv",
+                                                      "text/comma-separated-values,text/plain")),
+                                   title = "Upload rules",
+                                   content = "Upload the rules csv to use to validate the data csv"
+                               ) 
+                           },
+                           popover(
+                               downloadButton("download_rules_excel", "Data Template", style = "background-color: #ed6ca7;"),
+                               title = "Download rules template file",
+                               content = "This is a file that can be used as a template when collecting data so that it conforms to most of the rules tested in this portal."
+                           )
+                    ), 
+                    column(4, 
+                           uiOutput("alert"))),
+                    uiOutput("error_query"),
+                    uiOutput("dev_options")
+            ),
+            #tabItem(tabName = "item4",
[…]
