+++ inst/apps/data_visualization/README.md
@@ -1,0 +1,13 @@
+# Data Visualization Tool <a href="https://openanalysis.org/microplastic_analysis"><img src="man/DataVisualizationhex.png" alt="Data Visualization logo" align="right" style="height: 140px;"/></a>
+
+[![Website](https://img.shields.io/badge/web-openanalysis.org-white)](https://openanalysis.org/microplastic_analysis)
+
+# Description
+
+Welcome to the Data Visualization webpage. This tool displays visual characteristics and concentrations of microplastics through plots, tables, and maps. The first two tabs uses microplastics sample data from the California Open Data Portal and allows users to look at the concentration of microplastics by cities or countries. The last tab uses data that is entirely simulated for illustrative purposes to demonstrate the functionality of the application, including displaying yearly average concentrations, treatment levels, and polymer and width distributions.
+
+# Installation
+
+- Use the web app: [openanalysis.org/microplastic_analysis](https://openanalysis.org/microplastic_analysis)
+
+- The code for the app is in [/code/data_visualization](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/Microplastic_Data_Portal/tree/main/code/data_visualization) 
