+++ DESCRIPTION
@@ -72,5 +72,5 @@ Suggests:
     RCurl,
     purrr,
     stringr
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Config/testthat/edition: 3

[…]
