+++ inst/apps/code/microplastic_image_explorer/app.R
@@ -1,263 +1,0 @@
-library(shiny)
-library(shinyjs)
-library(bs4Dash)
-library(shinyWidgets)
-library(googlesheets4)
-library(dplyr)
-library(curl)
-
-file <- read.csv("image_metadata.csv")
-file$images <- paste0("https://d2jrxerjcsjhs7.cloudfront.net/", file$file_names)
-ui <- dashboardPage(
-  dashboardHeader(title = "Microplastic Image Explorer",
-                  fluidRow(
-                    column(
-                      width = 12,
-                      tags$div(
-                        tags$style(HTML("
-                      .navbar {
-                      background-color: #6D929B;
-                      }
-                      .breadcrumb {
-                      font-size: 14px;
-                      font-family: Arial, sans-serif;
-                      }"))
-                      ),
-                      tags$div(
-                        id = "breadcrumb",
-                        verbatimTextOutput("breadcrumb_output")
-                      )
-                    )
-                  )
-  ),
-  dashboardSidebar(
-    sidebarUserPanel(
-      #image = "https://drive.google.com/file/d/13iCjC10dV3giFhCCoir_8mnbwtHM1rMA/view?usp=sharing",
-      name = "Welcome!"
-    ),
-    tags$style(
-      HTML("
-      .sidebar { 
-      background-color: #6D929B;
-      }
-           ")
-    ),
-    sidebarMenu(
-      id = "sidebarmenu",
-      #sidebarHeader("Header 1"),
-      menuItem(
-        "Image Query",
-        tabName = "item1",
-        icon = icon("camera")
-      )
-    )
-  ),
-  dashboardBody(
-    tags$style(
-      HTML(".content-wrapper {
-              background-color: #F5FAFA;
-           }
-           .selected-border .selectize-input {
-              border-color: #C1DAD6;
-              border-width: 5px;
-              background-color: #C1DAD6;
-           }
-           .overview-box-content h3 {
-              font-size: 20px;
-           }
-           .contribute-box-content h3 {
-              font-size: 20px;
-           }
-        ")
-    ),
-    tags$script(HTML('
-    // JavaScript/jQuery to add custom class when a filter is selected
-    $(document).on("change", "select", function() {
-      $(this).parent().toggleClass("selected-border", $(this).val() !== "ALL");
-    });
-  ')),
-    tabItems(
-      tabItem(
-        tabName = "item1",
-        fluidRow(
-          box(
-            title = "Overview",
-            h3(
-              tags$div(
-                "Welcome to the microplastic taxonomy page!",
-                br(),
-                br(),
-                "This is a place to improve your use of visual microscopy in microplastic identification.",
-                br(),
-                br(),
-                "Go to the image query tab below to get started querying our database of microplastic images by color, morphology, and polymer types.",
-              )
-            ),
-            class = "overview-box-content",
-            width = 12
-          )
-        ),
-        fluidRow(
-          column(4,
-                 selectizeInput(inputId = "citation", 
-                                label = "Citation", 
-                                choices = c("ALL", toupper(unique(file$citation))),
-                                selected = "ALL"
-                 )
-          ),
-          column(4, 
-                 selectizeInput(inputId = "color", 
-                                label = "Color", 
-                                choices = c("ALL", toupper(unique(file$color))),
-                                selected = "ALL"
-                 )
-          ),
-          column(4,
-                 selectizeInput(inputId = "morphology", 
-                                label = "Morphology", 
[…]
