### 2024-05-10 cb6167e
Updates to package; fixing errors in R CMD check
