### 2024-05-10 f9e18b6
updates to vignettes; reduced image size by converting from png to jpg
