+++ vignettes/imageexplorer.Rmd
@@ -26,17 +26,17 @@ This document outlines a workflow to utilize the Microplastic Image Explorer fro
 To access the Image Explorer, go to this link [openanalysis.org/microplastic_image_explorer/](https://openanalysis.org/microplastic_image_explorer/) or go to our [Github](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/Microplastic_Data_Portal) and link it directly to your own device in R. After setting up the Github to your device, go to the `'microplastic_image_explorer'` folder and select the app.R file to run the app.
 
 ```{r, fig.align="center", echo=FALSE}
-knitr::include_graphics("appimages/ImageExplorerapp.png")
+knitr::include_graphics("appimages/ImageExplorerapp.jpg")
 ```
 
 Once the app has loaded, users can select between 'Citation', 'Color', 'Morphology', and 'Polymer' to query various microplastic images. The queried results will populate below, along with the breadcrumb filter at the top of the page which shows the filters that have been applied.
 
 ```{r, fig.align="center", echo=FALSE}
-knitr::include_graphics("appimages/ImageExplorerapp2.png")
+knitr::include_graphics("appimages/ImageExplorerapp2.jpg")
 ```
 
 The filter options will update with each filter applied to ensure that only existing images can be queried.
 
 ```{r, fig.align="center", echo=FALSE}
-knitr::include_graphics("appimages/ImageExplorerapp3.png")
+knitr::include_graphics("appimages/ImageExplorerapp3.jpg")
 ```
