### 2024-05-13 1f5446b
Added anonymous feedback form to 'Help' tab in validator; updated example_config.yml to include the feedback form (google form)
