### 2024-05-15 060d139
updates to paper.md (comments from coauthors)
