### 2024-05-16 c1738b0
separated valid keys so one is for sharing data and one is for downloading data
