### 2024-05-16 7dc2218
updated example_config for 'valid_key_share' and 'valid_key_download'
