### 2024-05-16 574acd9
updated readme and some of the functions for consistency throughout the package
