+++ R/global.R
@@ -29,8 +29,9 @@ certificate_df <- function(x, time = Sys.time()){
                                        Sys.info())))
 }
 
-#' Read rules from a file
+#' @title Read rules from a file
 #'
+#' @description
 #' This function reads rules from a file or a data frame.
 #' The file can be in csv or xlsx format.
 #' The data should have the column names "name", "description", "dataset", "valid example", "severity", "rule".
@@ -86,7 +87,7 @@ read_rules <- function(file_rules){
     return(rules)
 }
 
-#' Read and format data from csv or xlsx files
+#' @title Read and format data from csv or xlsx files
 #'
 #' @param files_data List of files to be read
 #' @param data_names Optional vector of names for the data frames
@@ -143,9 +144,9 @@ read_data <- function(files_data, data_names = NULL){
     data_formatted
 }
 
-#' Name datasets
+#' @title Name datasets
 #'
-#' This function extracts the names of the datasets provided in the input files.
+#' @description This function extracts the names of the datasets provided in the input files.
 #' If specific data names are provided, they are used, otherwise the function tries
 #' to extract the names from the files themselves.
 #'
@@ -175,9 +176,9 @@ name_data <- function(files_data, data_names = NULL){
 }
 
 
-#' Reformats the rules
+#' @title Reformat the rules
 #'
-#' This function is responsible for handling the rule reformating, dataset handling
+#' @description This function is responsible for handling the rule reformating, dataset handling
 #' and foreign key checks.
 #'
 #' @param rules A data.frame containing rules to be reformatted.
@@ -250,7 +251,7 @@ reformat_rules <- function(rules, data_formatted, zip_data = NULL){
 }
 
 
-#' Validate_data: Validate data based on specified rules.
+#' @title Validate data based on specified rules
 #'
 #' @param files_data A list of file paths for the datasets to be validated.
 #' @param data_names (Optional) A character vector of names for the datasets. If not provided, names will be extracted from the file paths.
@@ -401,10 +402,10 @@ validate_data <- function(files_data, data_names = NULL, file_rules = NULL, zip_
          issues = any_issues)
 }
 
-#' Remote Share Function
+#' @title Share your validated data
 #'
-#' This function uploads validated data to specified remote repositories,
-#' such as CKAN, Amazon S3, and/or MongoDB
+#' @description This function uploads validated data to specified remote repositories,
+#' such as CKAN, Amazon S3, and/or MongoDB.
 #'
 #' @param validation A list containing validation information.
 #' @param data_formatted A list containing formatted data.
@@ -554,9 +555,9 @@ remote_share <- function(validation, data_formatted, files, verified, valid_rule
                 submission_time = submission_time))
 }
 
-#' Download Structured Data from Remote Sources
+#' @title Download structured data from remote sources
 #'
-#' This function downloads data from remote sources like CKAN, AWS S3, and MongoDB.
+#' @description This function downloads data from remote sources like CKAN, AWS S3, and MongoDB.
 #' It retrieves the data based on the hashed_data identifier and assumes the data is stored using the same naming conventions provided in the `remote_share` function.
 #'
 #' @param hashed_data A character string representing the hashed identifier of the data to be downloaded.
@@ -650,9 +651,9 @@ remote_download <- function(hashed_data = NULL, ckan_url, ckan_key, ckan_package
     return(data_downloaded)
 }
 
-#' Download Raw Data from Remote Sources
+#' @title Download raw data from remote sources
 #'
-#' This function downloads data from remote sources like CKAN and AWS S3.
+#' @description This function downloads data from remote sources like CKAN and AWS S3.
 #' It retrieves the data based on the hashed_data identifier and assumes the data is stored using the same naming conventions provided in the `remote_share` function.
 #'
[…]
