+++ man/rows_for_rules.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/global.R
 \name{rows_for_rules}
 \alias{rows_for_rules}
-\title{rows_for_rules}
+\title{Check which rows in the data violated the rules}
 \usage{
 rows_for_rules(data_formatted, report, broken_rules, rows)
 }
