### 2024-05-16 ba38865
updated download_all function with an example
