+++ man/download_all.Rd
@@ -33,6 +33,19 @@ Any return objects from the downloads.
 This function allows users to download all data rather than one data set at a time.
 }
 \examples{
-# example code
+\dontrun{
+    download_all_data <- download_all(file_path = "your/path/file.zip",
+                                      s3_key_id = "your_s3_key_id",
+                                      s3_secret_key = "your_s3_secret_key",
+                                      s3_region = "your_s3_region",
+                                      s3_bucket = "your_s3_bucket",
+                                      callback = function(success) {
+                                      if (success) {
+                                      message("All files downloaded successfully.")
+                                      } else {
+                                      message("An error occurred during file download.")
+                                      }
+                                      })
+}
 
 }
