+++ R/global.R
@@ -745,13 +745,7 @@ remote_raw_download <- function(hashed_data = NULL, file_path = NULL, ckan_url =
 #'                                       s3_secret_key = "your_s3_secret_key",
 #'                                       s3_region = "your_s3_region",
 #'                                       s3_bucket = "your_s3_bucket",
-#'                                       callback = function(success) {
-#'                                       if (success) {
-#'                                       message("All files downloaded successfully.")
-#'                                       } else {
-#'                                       message("An error occurred during file download.")
-#'                                       }
-#'                                       })
+#'                                       callback = NULL)
 #' }
 #' 
 #' @export
