### 2024-05-16 5722e8f
updates to app.Rmd and package.Rmd
