### 2024-05-17 357ec2c
updates to readme for each of the apps
