+++ inst/apps/microplastic_image_explorer/README.md
@@ -10,7 +10,7 @@ Welcome to the Microplastic Image Explorer webpage. This tool was designed to al
 
 - Use the web app: [openanalysis.org/microplastic_image_explorer](https://openanalysis.org/microplastic_image_explorer/)
 
-- The code for the app is in [/inst/apps/microplastic_image_explorer](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/tree/main/inst/apps/microplastic_image_explorer) 
+- The code for the app is in [/inst/apps/microplastic_image_explorer](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/inst/apps/microplastic_image_explorer/app.R) 
 
 # Additional Information
 
