+++ inst/apps/validator/example_config.yml
@@ -18,10 +18,10 @@ default:
     # mongo_key: "your_mongo_key"
     # mongo_collection: "your_mongo_collection"
     # apiKey: "your_mongo_apikey"
-    rules_to_use: "www/microplastic_images/One4AllValidator_rules.csv"
-    rules_example: "www/microplastic_images/One4AllValidator_rules.csv"
-    valid_data_example: "www/microplastic_images/valid_example.xlsx"
-    invalid_data_example: "www/microplastic_images/invalid_example.xlsx"
+    rules_to_use: "www/rules_dw_acc.csv"
+    rules_example: "www/rules_dw_acc.csv"
+    valid_data_example: !expr c("www/data.xlsx", "www/extra_files.zip")
+    invalid_data_example: !expr c("www/data_example_errors.xlsx", "www/example_files_errors.zip")
     twitter: "https://twitter.com/Win_OpenData"
     github:  "https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/tree/main/inst/apps/validator"
     license: "https://opensource.org/license/mit"
