+++ paper/paper.md
@@ -38,6 +38,10 @@ authors:
   - name: Oluniyi O. Fadare
     orcid: 0000-0002-8712-0448
     affiliation: 2
+  - name: Anna Kukkola
+    affiliation: 8
+  - name: Hsuan-Cheng Lu
+    affiliation: 9
   - name: Win Cowger
     orcid: 0000-0001-9226-3104
     affiliation: "1, 2"
@@ -56,7 +60,11 @@ affiliations:
    index: 6
  - name: California State Water Resources Control Board, USA
    index: 7
-date: 15 April 2024
+ - name: University of Birmingham, Life and Environmental Sciences, UK
+   index: 8
+ - name: Australian Rivers Institute, School of Environment and Science, Griffith University, AUS
+   index: 9
+date: 20 May 2024
 bibliography: paper.bib
 
 ---
@@ -67,9 +75,9 @@ Microplastics are a diverse suite of contaminants [@Rochman:2019] requiring a va
 
 # Statement of need
 
-In the realm of scientific advancement, the ability to reuse data and reproduce results is paramount, as highlighted by @Peng:2011. For example, the National Center for Biotechnology Information (NCBI)'s gene databank contains over 3 billion records and is considered a critical component of biological research. Ample funding and attention has been directed toward the study of the human genome for decades, while plastic pollution, particularly microplastics, has only recently begun to receive this attention. The challenge in effectively sharing data, however, lies in the varied structuring and formatting of data across different research areas, especially in emerging fields such as plastic pollution. This inconsistency slows progress and results in inefficient use of resources to make disparate datasets compatible for analysis and decision making. For instance, the burgeoning field of microplastic pollution monitoring faces significant hurdles due to varied data collection methods, environmental matrices, and analytical approaches, leading to a lack of standardized data formats [@Hapich:2022; @Fox:2024; @Dissanayake:2022]. Although there are efforts towards creating databases, such as Čerkasova et al.'s public database for environmental microplastics [@Cerkasova:2023], widespread adoption of standardized data formats remains elusive.
+In the realm of scientific advancement, the ability to reuse data and reproduce results is paramount, as highlighted by @Peng:2011. For example, the National Center for Biotechnology Information (NCBI)'s gene databank contains over 3 billion records and is considered a critical component of biological research. Ample funding and attention has been directed toward the study of the human genome for decades, while plastic pollution, particularly microplastics, has only recently begun to receive this attention. The challenge in effectively sharing data, however, lies in the varied structuring and formatting of data across different research areas, especially in emerging fields such as plastic pollution. This inconsistency slows progress and results in inefficient use of resources to make disparate datasets compatible for analysis and decision making. For instance, the burgeoning field of microplastic pollution monitoring faces significant hurdles due to varied data collection methods, environmental matrices, and analytical approaches, leading to a lack of standardized data formats [@Lu:2021; @Hapich:2022; @Fox:2024; @Dissanayake:2022]. Although there are efforts towards creating databases, such as Čerkasova et al.'s public database for environmental microplastics [@Cerkasova:2023], widespread adoption of standardized data formats remains elusive.
 
[…]
