### 2024-05-20 7b7b9dd
updated validator tool link in readme
