### 2024-05-21 6d48a43
added pull_data function to check for duplicates; modified more png's to jpg's for package size
