+++ man/pull_data.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/global.R
+\name{pull_data}
+\alias{pull_data}
+\title{Check for duplicate entries}
+\usage{
+pull_data(mongo_collection, mongo_key, request = "{\\"samples.SampleID\\": 1}")
+}
+\arguments{
+\item{mongo_collection}{A character string representing the mongo collection.}
+
+\item{mongo_key}{A character string representing the mongo key.}
+
+\item{request}{A JSON string specifying the field to be retrieved. Modify to retrieve a different field.}
+}
+\value{
+A unique list of the requested data fields.
+}
+\description{
+This function checks for duplicate entries when uploading data. Data that has already been uploaded will be prevented if attempted to upload again.
+}
+\examples{
+\dontrun{
+mongo_key <- "your_mongo_key"
+mongo_collection <- "your_mongo_collection"
+data <- pull_data(mongo_collection, mongo_key, request = '{"samples.SampleID": 1}')
+print(data)
+}
+
+}
