### 2024-05-22 9dd7998
check for duplicates update (took out of package, added into shiny app)
