+++ NAMESPACE
@@ -10,7 +10,6 @@ export(create_valid_excel)
 export(download_all)
 export(is.POSIXct)
 export(name_data)
-export(pull_data)
 export(query_document_by_object_id)
 export(read_data)
 export(read_rules)
@@ -49,7 +48,6 @@ importFrom(dplyr,left_join)
 importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_if)
 importFrom(dplyr,select)
-importFrom(mongolite,mongo)
 importFrom(openxlsx,addWorksheet)
 importFrom(openxlsx,conditionalFormatting)
 importFrom(openxlsx,createStyle)
