### 2024-06-20 Moore-Institute-4-Plastic-Pollution-Res/One4All#104 @stephpenn1
runGithub() produces path error within run_app()
Hi @hsherrod2019,

I am trying to run `run_app(app = "validator")` in the console after installing the package via Github and am getting the error below (this happened for all three app calls). Debugging your function, the error happens in the runGithub functional call.

I'll note that the run_app call works on the CRAN package version for the validator app, so this may be due to any recent changes you've made with integrating the apps into one package.

I appreciate any guidance, thanks!

```
remotes::install_github("Moore-Institute-4-Plastic-Pollution-Res/One4All")
library(One4All)
run_app(app = "validator")
Downloading https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/archive/main.tar.gz
Error in `shinyAppDir()`:
! No Shiny application exists at the path "/var/folders/fh/_qqkjwtj50106y9tr_jb_rhh0000gn/T//RtmpU8vdcs/shinyapp58e5d1886ce/One4All-main//inst/apps/code/data_visualization"
Run `rlang::last_trace()` to see where the error occurred.
```

```
rlang::last_trace()
<error/invalidShinyAppDir>
Error in `shinyAppDir()`:
! No Shiny application exists at the path "/var/folders/fh/_qqkjwtj50106y9tr_jb_rhh0000gn/T//RtmpH9WAlB/shinyapp30126182a10b/One4All-main//inst/apps/code/validator"
---
Backtrace:
    ▆
 1. └─One4All::run_app(app = "validator")
 2.   └─shiny::runGitHub(...)
 3.     └─shiny::runUrl(url, subdir = subdir, destdir = destdir, ...)
 4.       └─shiny::runApp(appdir, ...)
 5.         ├─shiny::as.shiny.appobj(appDir)
 6.         └─shiny:::as.shiny.appobj.character(appDir)
 7.           └─shiny::shinyAppDir(x)
Run rlang::last_trace(drop = FALSE) to see 1 hidden frame.
```

_This Github issue is related to openjournals/joss-reviews#6715_
