### 2024-06-20 Moore-Institute-4-Plastic-Pollution-Res/One4All#104 @stephpenn1
All 3 apps run now, thank you!
