### 2024-06-25 #6715 @hahsan1
I have successfully installed the package and currently going through the documentation. In general the documentation is very good and has useful technical content.

I am now following the tutorial. I attempted to run the validator app but ran into an issue. I downloaded rules.csv and valid_data.zip and have uploaded both, but get this error:

![image](https://github.com/openjournals/joss-reviews/assets/72406931/5bf677b5-5d0c-4642-9a9f-a681122c2f11)

I noticed that the excel file in the zip folder has three tabs, with the tab labeled "particles" having only headers but no data underneath it. Is this the issue?

Also, my understanding is that the validator app is supposed to read a file with rules to match against a dataset to see if the entries in each data column are formatted correctly and/or contain the right kind of data? Perhaps that could be more explicitly mentioned in the guide, unless I missed it.
