### 2024-06-25 #6715 @hsherrod2019
> hahsan1

@hahsan1 , thank you for letting me know! The rules file should be 'rules_dw_acc.csv' and is already implemented into the tool. When you upload the zip folder, I recommend extracting the 'data.xlsx' (structured data) and the 'extra_files.zip' (unstructured data) and uploading those two together rather than the zip folder itself.

The data.xlsx should have the three sheets (methodology, samples, particles) with data in each of those sheets. Did you download the sample data from the tool itself? I can further investigate into this!

Yes, the tool and package have a rules sheet applied ahead of time so that when a user uploads their data, the tool/package functions can 'validate' the data and check to see if it's harmonized:)
