### 2024-06-25 #6715 @hsherrod2019
@hahsan1 , I've attached a screenshot below. When you download the valid sample data, if you click on the zip folder the two files should appear. Let me know if that doesn't work! Also the rules file defaults to saving it as 'rules.csv' in the user's local device, but it is the 'rules_dw_acc.csv' from github:)

![Screenshot 2024-06-25 at 10 30 38 AM](https://github.com/openjournals/joss-reviews/assets/135265754/c4bed9e1-9bde-41ca-ab27-f2b5e36a2a6c)
