### 2024-06-25 #6715 @hahsan1
The valid sample zip file only has valid_example.xlsx. (By the way, I had installed the release version of this package from CRAN, maybe these other files are on the development version?)
