### 2024-06-26 #6715 @hsherrod2019
@hahsan1 , I just downloaded it from the CRAN version and I'm getting the same error as well (valid_example.xlsx). I also downloaded the github version and that is working so I recommend downloading the package through github. Please let me know if you are still facing the 'rules_broken' error through the github version. I will look into it in the meantime as well!
