### 2024-06-26 #6715 @stephpenn1
Hi, @hsherrod2019 and @kanishkan91, I wanted to confirm before I complete my review that the paper should only be describing the validator app. I know that the other two apps have been integrated into the package but are listed in the Future Goals section of the paper. Thanks!
