### 2024-06-26 #6715 @hsherrod2019
> > @hahsan1 , I just downloaded it from the CRAN version and I'm getting the same error as well (with only valid_example.xlsx). I also downloaded the github version and that is working so I recommend downloading the package through github. Please let me know if you are still facing the 'rules_broken' error through the github version. I will look into it in the meantime as well!
>
> I installed the github version and now get the files you mentioned. So I uploaded the data.xlsx and extra_files.zip, and it gives me a warning and an error under "samples". The error is "Checking for duplicate entries", but when I select it, it doesn't show up on the issues selected window. A message pops up saying "Column being assessed by the rule is not in the dataset". Is this an expected behavior?
>
> Also, I did try the `rules_broken` function, but still get the same error. It seems like it's not compatible with the list output from the `validate_data` function.

Hi @hahsan1 , I'm currently working through these errors and will get back to you as soon as I can! Thank you!
