### 2024-06-28 #6715 @hahsan1
Hi @kanishkan91 and @hsherrod2019 , I want to note that I will be on vacation for three weeks starting Monday, so I want to provide some final comments so that I don't hold up the review. I've had a chance to look at all three tools. Overall the tools are very well made and generally easy to use. It seems that the validator app is quite versatile, if it allows the user to create their own rules sheet to validate data against. The other tools are specific for microplastic visualization and identification. A few questions/comments about the tools.

Validator: Is there a way to upload a custom rules sheet? I might have missed this in the tutorial. I believe you mentioned that there is a rules sheet already embedded in the tool.

Data visualization tool: The map isn’t populated with the microplastics data until a country is selected in the search bar. Would be nice to have them all there by default.

Microplastic Image Explorer: Just a minor comment. I found that clicking on an image opens an empty window.
