### 2024-07-01 #6715 @hsherrod2019
> Hi @kanishkan91 and @hsherrod2019 , I want to note that I will be on vacation for three weeks starting Monday, so I want to provide some final comments so that I don't hold up the review. I've had a chance to look at all three tools. Overall the tools are very well made and generally easy to use. It seems that the validator app is quite versatile, if it allows the user to create their own rules sheet to validate data against. The other tools are specific for microplastic visualization and identification. A few questions/comments about the tools.
>
> Validator: Is there a way to upload a custom rules sheet? I might have missed this in the tutorial. I believe you mentioned that there is a rules sheet already embedded in the tool.
>
> Data visualization tool: The map isn’t populated with the microplastics data until a country is selected in the search bar. Would be nice to have them all there by default.
>
> Microplastic Image Explorer: Just a minor comment. I found that clicking on an image opens an empty window.

Hi @hahsan1 and @kanishkan91 ,

Thank you so much for the feedback! The responses are below:

Validator: Yes, users can upload a custom rules sheet. I've added additional instructions on how to do this in the tutorial section of the paper. Users can define which rules sheet to use in the configuration file (field 'rules_to_use'), or they can hash out the field and upload the rules sheet directly in the app.

Data Visualization Tool: I just integrated those changes, the locations should now appear by default.

Image Explorer: Unfortunately, that appears to be an issue with the R browser. Clicking the image opens another browser and I don't think R itself is able to handle this, which is why it appears blank. However, if you open the tool in the web browser, this shouldn't be an issue!

Also, the 'check for duplicates' error is now resolved as well! Thank you.
