### 2024-07-01 c655449
updates to data visualization tool; locations appear by default
