+++ inst/apps/data_visualization/app.R
@@ -209,34 +209,37 @@ ui <- bs4DashPage(
         tabName = "MapByRawData",
         fluidRow(
           box(
-            title = "About",
+            title = "Welcome to the Data Visualization Tool!",
             h3(
               tags$div(
-                "Welcome to the data visualization portal!",
-                br(),
-                br(),
-                "This app uses microplastics sample data from the California Open Data Portal.",
+                "This app uses microplastics sample data from the California (CA) Open Data Portal.",
                 br(),
                 br(),
                 "Below is a map that allows you to select multiple locations, and the popup markers present overview information about the plastics at those sites.",
                 br(),
                 br(),
-                "The concentration and source of plastics of the selected locations are then displayed in a rain cloud plot below.",
+                "The concentration and source of plastics of the selected locations are displayed in a rain cloud plot below.",
                 br(),
                 br(),
-                "The table directly generates the data from the Open Data Portal, and shows additional information regarding the characteristics of the plastics.",
+                "The table directly generates data from the CA Open Data Portal and shows additional information regarding the characteristics of the plastics.",
                 br(),
                 br(),
-                "Warning: The latitude and longitude data for some locations are inaccurate due to multiple locations being grouped together within a single sample."
+                "Warning: The latitude and longitude data for some locations are inaccurate due to multiple locations being grouped together within a single sample.",
+                style = "font-size: 14px;"
               )
             ),
             class = "about-box-content",
             width = 12
             ),
           column(
-            width = 12,
-            selectInput("Location", "Location", choices = Location_choices, multiple = TRUE)
-          )
+              width = 6,
+              selectInput("Location", "Location", choices = Location_choices, multiple = TRUE)
+              ),
+          column(
+              width = 6,
+              align = "right",
+              actionButton(inputId = "clearAllButtonLocation", "Clear All")
+            )
         ),
         fluidRow(
           column(
@@ -266,16 +269,32 @@ ui <- bs4DashPage(
             DT::dataTableOutput("tableLocation"), 
             width = 12
           )
-        )
-      ),
+          )
+        ),
       tabItem(
         tabName = "MapByCountries",
         fluidRow(
-          column(
-            width = 12,
+            box(
+                title = "Map By Countries",
+                h3(
+                    tags$div(
+                        "Use this tab to filter by country.",
+                        style = "font-size: 14px;"
+                        )
+                    ),
+                class = "about-box-content",
+                width = 12
+                ),
+            column(
+            width = 6,
             selectInput("Country", "Country", choices = Country_choices, multiple = TRUE)
-          )
-        ),
+            ),
+            column(
+                width = 6,
+                align = "right",
+                actionButton(inputId = "clearAllButtonCountry", "Clear All")
+                )
+            ),
         fluidRow(
           column(
             width = 12,
@@ -316,7 +335,8 @@ ui <- bs4DashPage(
                 "Sankey plots are a useful way to visualize the relationship and flow between variables.",
                 br(),
                 br(),
-                "The drop down allows users to look at the relationships of three different microplastic characteristics from the Open Data Portal."
+                "The drop down allows users to look at the relationships of three different microplastic characteristics from the CA Open Data Portal.",
+                style = "font-size: 14px;"
[…]
