### 2024-07-01 #6715 @hsherrod2019
> > > > Hi, @hsherrod2019 and @kanishkan91, I wanted to confirm before I complete my review that the paper should only be describing the validator app. I know that the other two apps have been integrated into the package but are listed in the Future Goals section of the paper. Thanks!
> > >
> > >
> > > Hi @stephpenn1 , the paper is mainly focused on the validator tool and package, with the exception of the other two tools in the Future Goals section. However, the pkgdown (https://moore-institute-4-plastic-pollution-res.github.io/One4All/) has vignettes for each of the tools + package which we were hoping to be reviewed as well. Thank you!
> >
> >
> > @hsherrod2019 Quick clarifying question on this. Based on our discussion earlier, since these new applications are now integrated into the same repository, the paper would need to be updated to reflect the same rather than mentioned as "future directions", correct? If so, I would recommend you make the requisite additions to the paper.
>
> Hi @kanishkan91 , the two tools had been added to the flowchart, and the Microplastics Data Portal is discussed in the Statement of Need. I will also add the two tools into the tutorial section to explain how they are related to the validator tool. Thank you for your patience!

Hi @kanishkan91 I'm currently working on integrating these changes to the paper and will update it on github (along with the custom rules sheet update) as soon as possible. Thank you.
