### 2024-07-01 c8e3b18
update to main readme and image explorer readme to reflect updates to paper
