### 2024-07-02 999471a
update to image explorer font size
