### 2024-07-02 adeede0
update to manuscript; incorporated data visualization and image explorer in main text, modified future goals section
