+++ paper/paper.md
@@ -68,7 +68,7 @@ affiliations:
    index: 9
  - name: Syracuse University, USA
    index: 10
-date: 20 May 2024
+date: 2 July 2024
 bibliography: paper.bib
 
 ---
@@ -91,15 +91,13 @@ A full tutorial video is available on [YouTube](https://www.youtube.com/embed/LM
 
 ## Workflow Overview
 
-Consisting of an R package and an R shiny application, this portal was designed to enhance data validation and management processes by employing a set of functions that read a set of rules from a ‘CSV’ or ‘Excel’ file to validate a dataset. Users have the option to work in the validator app or the One4All package. The functionality in the validator app is based on the One4All package (Figure 1). 
+Consisting of an R package and R shiny application, this tool was designed to enhance data validation and management processes by employing a set of functions that read a set of rules from a ‘CSV’ or ‘Excel’ file to validate a dataset. Users have the option to work in the Validator app or the One4All package. The functionality in the application is based on the One4All package (Figure 1). 
 
-![Workflow of the One4All portal. Navigate through the One4All package and the validator app, based on user preference: (https://lucid.app/documents/view/f33867e8-5822-4cef-b62a-3a6c2e293b4e).\label{fig:example1}](manuscriptimages/One4AllFlowChart.png)
+![Workflow of the One4All portal. Navigate through the One4All package and the Validator app, based on user preference: (https://lucid.app/documents/view/f33867e8-5822-4cef-b62a-3a6c2e293b4e).\label{fig:example1}](manuscriptimages/One4AllFlowChart.png)
 
-The set of rules comes from a [rules sheet](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/inst/apps/validator/www/rules_dw_acc.csv), which is a 'CSV' file that contains the following fields: name, description, dataset, valid example, severity, and rule. Users should add the rules as R functions, referring to the column names in the uploaded data. The dataset is an optional field when the file is separated into multiple sheets or files (Figure 2).
+The rules sheet and configuration file power the application and package functions to validate, share, and download data. The implemented set of rules comes from a [rules sheet](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/blob/main/inst/apps/validator/www/rules_dw_acc.csv), which is a 'CSV' file containing the following fields: name, description, dataset, valid example, severity, and rule. Users can adapt the rules sheet to use the tool for their own purposes. Add the rules as R functions, referring to the column names in the uploaded data. The dataset is an optional field when the file is separated into multiple sheets or files. Specify the rules sheet to use in the 'rules_to_use' configuration field, or hash out the field and upload the rules sheet directly within the app (Figure 2).
 
-![A screenshot showing a subset of the example rules sheet applied in the One4All portal, which comprises a total of 82 rules. The validator app provides users with sample rules, a valid data example, and an invalid data example, all of which are available for users to download and adapt to suit their own purposes.\label{fig:example2}](manuscriptimages/One4AllRulesSheet.png)
-
-The rules sheet and configuration file (explained below) power the application and package functions to validate, share, and download data. 
+![A screenshot showing a subset of the example rules sheet applied in the One4All portal, which comprises a total of 81 rules. The Validator app provides users with sample rules, a valid data example, and an invalid data example, all of which are available for users to download and adapt to suit their own purposes.\label{fig:example2}](manuscriptimages/One4AllRulesSheet.png)
 
[…]
