### 2024-07-02 2752f19
Update paper.md
Halfway
