### 2024-07-02 5fab998
update to package vignette with detailed example of 'rules_broken' function
