### 2024-07-02 1239ead
added data.validator citation to paper.bib
