+++ paper/paper.bib
@@ -294,3 +294,11 @@ API},
     year = {2021},
     doi = {10.1016/j.scitotenv.2021.146693}
 }
+
+@Manual{Appsilon:2024,
+  title = {data.validator: Automatic Data Validation and Reporting},
+  author = {Marcin Dubel and Paweł Przytuła and Jakub Nowicki and Krystian Igras},
+  year = {2024},
+  note = {R package version 0.2.1, https://github.com/Appsilon/data.validator},
+  url = {https://appsilon.github.io/data.validator/},
+}
