### 2024-07-02 54f5d7a
syntax fixes + added data validator citation in paper.md
