+++ paper/paper.md
@@ -244,7 +244,7 @@ The main limitation is finding a balance between expanding the variety of rules 
 
 # Availability and Documentation:
 
-One4All is published on CRAN and was created in R(4.2.1) [@R:2022] using the following libraries: shiny [@Shiny:2023], dplyr [@Dplyr:2023], validate [@Validate:2021], digest [@Digest:2023], data.table [@Datatable:2023], ckanr [@Ckanr:2023], openxlsx [@Openxlsx:2023], lexicon [@Lexicon:2018], readr [@Readr:2023], readxl [@Readxl:2023], tibble [@Tibble:2023], aws.s3 [@Awss3:2020], rlang [@Rlang:2023], jsonlite [@Jsonlite:2014], mongolite [@Jsonlite:2014], and httr [@Httr:2023]. The microplastic version of the validator app is hosted on the web at [openanalysis.org/one4all/](https://openanalysis.org/microplastic_data_portal/). The source code is [available on GitHub](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All) with an [MIT License](https://opensource.org/license/mit). The Moore Institute for Plastic Pollution Research maintains the package. The source code can be run offline on any machine that can install the One4All package. 
+One4All is published on CRAN and was created in R(4.2.1) [@R:2022] using the following libraries: shiny [@Shiny:2023], dplyr [@Dplyr:2023], validate [@Validate:2021], digest [@Digest:2023], data.table [@Datatable:2023], ckanr [@Ckanr:2023], openxlsx [@Openxlsx:2023], lexicon [@Lexicon:2018], readr [@Readr:2023], readxl [@Readxl:2023], tibble [@Tibble:2023], aws.s3 [@Awss3:2020], rlang [@Rlang:2023], jsonlite [@Jsonlite:2014], mongolite [@Jsonlite:2014], and httr [@Httr:2023]. The microplastic version of the Validator app is hosted on the web at [openanalysis.org/one4all/](https://openanalysis.org/microplastic_data_portal/). The source code is [available on GitHub](https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All) with an [MIT License](https://opensource.org/license/mit). The Moore Institute for Plastic Pollution Research maintains the package. The source code can be run offline on any machine that can install the One4All package. 
 
 # Future Goals
 
@@ -252,6 +252,6 @@ The One4All portal is intended to be useful for users to validate and share thei
 
 # Acknowledgments
 
-We acknowledge and greatly appreciate the financial support from the National Renewable Energy Laboratory and the Possibility Lab. The Moore Institute for Plastic Pollution Research led the development of the software tool and drafting of this manuscript. The One4All portal was inspired by the data.validator tool created by [@Appsilon:2024]. We acknowledge the work and the input of the San Francisco Estuary Institute and the California State Water Resources Control Board. We would like to thank Scott Coffin, Tony Hale, Diana Lin, Gemma Shusterman, Rebecca Sutton, Adam Wong, Richard Nelson, Leah Thornton Hampton, Libby Heeren, and Gabriel Daiess. The views and opinions expressed in this article are those of the authors and do not necessarily reflect the official policy or position of any California State agency.
[…]
