+++ NEWS.md
@@ -1,5 +1,5 @@
 # One4All 0.1
 
-## Innitial Release!
+## Initial Release!
 
 - Just getting started, lots of work left to do. 
