### 2024-07-12 #6715 @stephpenn1
Hi @hsherrod2019 and @kanishkan91 , I have completed my review and have summarized comments below, including any new questions and those already addressed. Overall the package is very well documented and the UI is smooth! Nicely prepared and this was a pleasure to review.

_Paper:_
- Thank you for updating the paper to include the other two apps within One4All, this new wording more accurately reflects your software. I think the paper reads well and looks good.

_Software Functionality:_
- There was a minor bug in the run_app function (https://github.com/Moore-Institute-4-Plastic-Pollution-Res/One4All/issues/104) which has now been resolved and works as described in the paper.
- When installing and running the package for the first time, I noticed you have a handy error message listing the packages that need to be installed. This is super optional but one suggestion would be to list the packages in an error messafe that could be copied directly into the console (e.g. `Please use install.packages(c(package1, package2))`)
- The data already integrated into your app - does this come from your own plastics data repository? And the validator app is for users looking to upload their own data?

_Documentation:_
- Now that you have made some changes to the functions, do you plan on resubmitting to CRAN? If so, documentation is good as is. If not, it should be clear this paper focuses on the github version.
