### 2024-07-12 #6715 @hsherrod2019
Hi @stephpenn1 ,

Thank you so much for the comments!

_Software Functionality:_
- Yes, the data already integrated comes from our own plastics data repository ('data.xlsx' and 'extra_files.zip'). Also yes, the validator app is for users looking to upload their own data.

- I will update the error message to include the packages, early next week, and will let you and @kanishkan91 know once I have integrated those changes!

_Documentation:_
- I resubmitted to CRAN last week after implementing changes to the package. The changes were approved and everything is up to date on CRAN now as well!

Thank you again.
