### 2024-07-13 #6715 @kanishkan91
👋 @hsherrod2019  - Once you address the comment above as you said and the repository is finalized, we need to set up the archive for your new release.

So after the last issues are resolved:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

- [ ] Check that software archive exists and title and author list match JOSS paper (or purposefully do not).

 Please respond with the 1) version and 2) DOI of the archived version here
