### 2024-07-15 1aaf497
update to run_app function; updated error message so users can copy missing packages to download from console
