### 2024-07-15 #6715 @hsherrod2019
Hi @kanishkan91 ,

I have resolved the last issues!

1) 2.0
2) https://doi.org/10.5281/zenodo.12747059
