### 2024-07-15 #6715 @kanishkan91
@hsherrod2019  Thanks ! I have recommended this for acceptance now . The AEiC in this submission track will review shortly and if all goes well this will go live soon! Before I forget, big thank you to @stephpenn1 and @hahsan1 for reviewing this paper! Your expertise and insights have been invaluable to this process. JOSS relies heavily on reviewers such as yourself. I will leave a positive review for your reviewing efforts on the JOSS reviewer hub. (PS- The positive reviews may lead to more invitations to review for JOSS in the future! My apologies in advance)

@hsherrod2019  - I will now go through the paper and post some comments if I find any typos or anything.
