### 2024-07-15 #6715 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
