### 2024-07-16 #6715 @crvernon
@hsherrod2019 - Things are looking good, I just need you to address the following before I accept this one:

- [ ] Some of the headers in the paper have a colon behind them.  Please remove the colons to make things consistent throughout the paper.
- [ ] Please ensure items needing capitalization in the references in your bib file are maintained after the paper is compiled.  You can do this by adding curly brackets around items needing to maintain formatting.  For example, LINE 186 should have "r" capitalized.  Please check to make sure there are not others that need to be fixed as well.

No need to conduct another release.  Once these have been changed, I will move forward with acceptance.
