### 2024-07-16 d143d1b
edits to paper.bib capitalization
