+++ paper/paper.bib
@@ -32,7 +32,7 @@
 
 @article{Cowger:2021,
     author = {Cowger, W. and Steinmetz, Z. and Gray, A. and Munno, K. and Lynch, J. and Hapich, H. and Primpke, S. and DeFrond, H. and Rochman, C. and Herodotou, O.},
-    title = {Microplastic Spectral Classification Needs an Open Source Community: Open Specy to the Rescue!},
+    title = {Microplastic Spectral Classification Needs an Open Source Community: {Open Specy} to the Rescue!},
     journal = {Analytical Chemistry},
     volume = {93},
     pages = {7543-7548},
@@ -51,7 +51,7 @@
 
 @article{Cowger:2023,
     author = {Cowger, W. and Gomez, I. and Martinez-Rubin, N. and Moriarity, A. and Harwell, T. and Anich, L.},
-    title = {Community Science-Informed Local Policy: A Case Study in Pinole Creek Litter Assessment},
+    title = {Community Science-Informed Local Policy: A Case Study in {Pinole Creek} Litter Assessment},
     journal = {Community Science},
     volume = {2},
     number = {1},
@@ -76,7 +76,7 @@
 
 @article{Hapich:2022,
     author = {Hapich, H. and Cowger, W. and Gray, A. and Tangri, N. and Hale, T. and Magdy, A. and Vermilye, A. and Yu, W. and Ayres, D. and Moore, C. and Vermilye, J. and Singh, S. and Haiman, A. N. K. and Youngblood, K. and Kang, Y. and McCauley, M. and Lok, T. and Moore, S. and Baggs, E. and Lippiatt, S. and Kohler, P. and Conley, G. and Taing, J. and Mock, J.},
-    title = {Trash Taxonomy Tool: harmonizing classification systems used to describe trash in environments},
+    title = {{Trash Taxonomy Tool}: harmonizing classification systems used to describe trash in environments},
     journal = {Microplastics and Nanoplastics},
     volume = {2},
     pages = {15},
@@ -133,7 +133,7 @@
 }
 
 @Manual{Shiny:2023,
-  title = {shiny: Web Application Framework for R},
+  title = {shiny: Web Application Framework for {R}},
   author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
   year = {2023},
   note = {R package version 1.8.0},
@@ -160,7 +160,7 @@
 }
 
 @Manual{Digest:2023,
-  title = {digest: Create Compact Hash Digests of R Objects},
+  title = {digest: Create Compact Hash Digests of {R} Objects},
   author = {Dirk Eddelbuettel with contributions by Antoine Lucas and Jarek Tuszynski and Henrik Bengtsson and Simon Urbanek and Mario Frasca and Bryan Lewis and Murray Stokely and Hannes Muehleisen and Duncan Murdoch and Jim Hester and Wush Wu and Qiang Kou and Thierry Onkelinx and Michel Lang and Viliam Simko and Kurt Hornik and Radford Neal and Kendon Bell and Matthew {de Queljoe} and Ion Suruceanu and Bill Denney and Dirk Schumacher and Winston Chang and Dean Attali.},
   year = {2023},
   note = {R package version 0.6.33},
@@ -233,7 +233,7 @@ API},
 }
 
 @Manual{Rlang:2023,
-  title = {rlang: Functions for Base Types and Core R and 'Tidyverse' Features},
+  title = {rlang: Functions for Base Types and Core {R} and 'Tidyverse' Features},
   author = {Lionel Henry and Hadley Wickham},
[…]
