### 2024-07-16 57e8196
updates to paper.bib capitalization; update to paper.md colon heading
