+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Cerkasova:2023,
     author = {Čerkasova, N. and Enders, K. and Lenz, R. and Oberbeckmann, S. and Brandt, J. and Fischer, D. and Fischer, F. and Labrenz, M. and Schernewski, G.},
-    title = {A Public Database for Microplastics in the Environment},
+    title = {{A Public Database for Microplastics in the Environment}},
     journal = {Microplastics},
     volume = {2},
     number = {1},
@@ -11,7 +11,7 @@
 
 @article{Coffin:2023,
     author = {Coffin, S.},
-    title = {The emergence of microplastics: Charting the path from research to regulations},
+    title = {{The emergence of microplastics: charting the path from research to regulations}},
     journal = {Environmental Science: Advances},
     volume = {2},
     number = {3},
@@ -32,7 +32,7 @@
 
 @article{Cowger:2021,
     author = {Cowger, W. and Steinmetz, Z. and Gray, A. and Munno, K. and Lynch, J. and Hapich, H. and Primpke, S. and DeFrond, H. and Rochman, C. and Herodotou, O.},
-    title = {Microplastic Spectral Classification Needs an Open Source Community: {Open Specy} to the Rescue!},
+    title = {{Microplastic Spectral Classification Needs an Open Source Community: Open Specy to the Rescue!}},
     journal = {Analytical Chemistry},
     volume = {93},
     pages = {7543-7548},
@@ -51,7 +51,7 @@
 
 @article{Cowger:2023,
     author = {Cowger, W. and Gomez, I. and Martinez-Rubin, N. and Moriarity, A. and Harwell, T. and Anich, L.},
-    title = {Community Science-Informed Local Policy: A Case Study in {Pinole Creek} Litter Assessment},
+    title = {{Community Science-Informed Local Policy: A Case Study in Pinole Creek Litter Assessment}},
     journal = {Community Science},
     volume = {2},
     number = {1},
@@ -61,7 +61,7 @@
 
 @book{California:2022,
     author = {{Division of Drinking Water, State Water Resources Control Board, State of California}},
-    title = {Policy Handbook Establishing a Standard Method of Testing and Reporting of Microplastics in Drinking Water},
+    title = {{Policy Handbook Establishing a Standard Method of Testing and Reporting of Microplastics in Drinking Water}},
     year = {2022}
 }
 
@@ -76,7 +76,7 @@
 
 @article{Hapich:2022,
     author = {Hapich, H. and Cowger, W. and Gray, A. and Tangri, N. and Hale, T. and Magdy, A. and Vermilye, A. and Yu, W. and Ayres, D. and Moore, C. and Vermilye, J. and Singh, S. and Haiman, A. N. K. and Youngblood, K. and Kang, Y. and McCauley, M. and Lok, T. and Moore, S. and Baggs, E. and Lippiatt, S. and Kohler, P. and Conley, G. and Taing, J. and Mock, J.},
-    title = {{Trash Taxonomy Tool}: harmonizing classification systems used to describe trash in environments},
+    title = {{Trash Taxonomy Tool: harmonizing classification systems used to describe trash in environments}},
     journal = {Microplastics and Nanoplastics},
     volume = {2},
     pages = {15},
@@ -86,7 +86,7 @@
 
 @article{Jenkins:2022,
[…]
