### 2024-07-16 #6715 @hsherrod2019
@crvernon Thank you!

I have removed the colons from the headers and ensured capitalization for the references. Both changes are pushed to GitHub now as well.
