### 2024-07-16 #6715 @crvernon
@xuanxu I am getting a 404 for this paper display when I follow provided link. Are you seeing this as well? Or is this just on my end?
